+++ ui/include/rawrbox/ui/elements/label.hpp
@@ -50,7 +50,7 @@ namespace rawrbox {
 		[[nodiscard]] virtual const std::string& getText() const;
 
 		virtual void setFont(rawrbox::Font* font);
-		virtual void setFont(const std::filesystem::path& font, int size = 11);
+		virtual void setFont(const std::filesystem::path& font, uint16_t size = 11);
 		[[nodiscard]] virtual rawrbox::Font* getFont() const;
 
 		virtual void sizeToContents();
