+++ ui/include/rawrbox/ui/scripting/wrappers/elements/label_wrapper.hpp
@@ -26,7 +26,7 @@ namespace rawrbox {
 		[[nodiscard]] virtual const std::string& getText() const;
 
 		virtual void setFont(const rawrbox::FontWrapper& font);
-		virtual void setFont(const std::string& font, sol::optional<int> size, sol::this_environment modEnv);
+		virtual void setFont(const std::string& font, sol::optional<uint16_t> size, sol::this_environment modEnv);
 		[[nodiscard]] virtual rawrbox::FontWrapper getFont() const;
 
 		virtual void sizeToContents();
